Adding Comments, did some QA testing, fixed bugs when array variable wasn't updated when deleting tasks or when the newely added clearALL function is used.
Adding Comments, did some QA testing, fixed bugs when array variable wasn't updated when deleting tasks or when the newely added clearALL function is used.

There's amillion ways of doing this task list exercise, I wanted to see if I could get it done on my own entirely, a bit of a mess, lots of bugs I had to fix, but it was a good learning exercise.